moved validators to new folder and updated namespace and use statements
